# docs/openapi/external_probe_api.yaml
openapi: 3.1.0
info:
  title: SSNP External Probe Ingestion API
  version: 0.1.0
  summary: Probe-side ingestion contract for SSNP qualification evidence
  description: |
    Minimal v0.1 contract for external probe evidence used by the Qualification
    Engine. Raw probe events are ingested first and daily qualification summaries
    are derived from those immutable records.
servers:
  - url: https://ssnp.example.org
    description: Example production endpoint
...
paths:
...
  /api/v1/probes/daily-summaries/{node_id}/{date_utc}:
    get:
      tags: [Probe]
      summary: Read one derived daily qualification summary
      operationId: getDailyQualificationSummary
      parameters:
        - name: node_id
          in: path
          required: true
          schema:
            type: string
        - name: date_utc
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Daily summary returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyQualificationSummary"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "429":
          $ref: "#/components/responses/ErrorResponse"
components:
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    DailyQualificationSummary:
      type: object
      additionalProperties: false
      required:
        - node_id
        - date_utc
        - policy_version
        - finalized_lag_threshold_blocks
        - chain_lag_threshold_blocks
        - valid_probe_count
        - availability_up_count
        - availability_ratio
        - finalized_lag_measurable_count
        - finalized_lag_within_threshold_count
        - finalized_lag_ratio
        - chain_lag_measurable_count
        - chain_lag_within_threshold_count
        - chain_lag_ratio
        - region_count
        - availability_passed
        - finalized_lag_passed
        - chain_lag_passed
        - multi_region_evidence_passed
        - qualified_probe_evidence_passed
      properties:
        node_id:
          type: string
        date_utc:
          type: string
          format: date
        policy_version:
          type: string
        finalized_lag_threshold_blocks:
          type: integer
          minimum: 1
        chain_lag_threshold_blocks:
          type: integer
          minimum: 1
        valid_probe_count:
          type: integer
          minimum: 0
        availability_up_count:
          type: integer
          minimum: 0
        availability_ratio:
          type: number
          minimum: 0
          maximum: 1
        finalized_lag_measurable_count:
          type: integer
          minimum: 0
        finalized_lag_within_threshold_count:
          type: integer
          minimum: 0
        finalized_lag_ratio:
          type: number
          minimum: 0
          maximum: 1
        chain_lag_measurable_count:
          type: integer
          minimum: 0
        chain_lag_within_threshold_count:
          type: integer
          minimum: 0
        chain_lag_ratio:
          type: number
          minimum: 0
          maximum: 1
        region_count:
          type: integer
          minimum: 0
        availability_passed:
          type: boolean
        finalized_lag_passed:
          type: boolean
        chain_lag_passed:
          type: boolean
        multi_region_evidence_passed:
          type: boolean
        qualified_probe_evidence_passed:
          type: boolean
        insufficient_evidence_reason:
          type:
            - string
            - "null"
        generated_at:
          type:
            - string
            - "null"
          format: date-time
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - status
        - error_code
        - message
      properties:
        status:
          type: string
          const: error
        error_code:
          type: string
          enum:
            - invalid_schema_version
            - missing_required_field
            - invalid_lag_value
            - duplicate_probe_id
            - unknown_node_id
            - unknown_region_id
            - stale_timestamp
            - rate_limited
        message:
          type: string